(slightly) faster imgui rendering.
In the past, I've found 'np.ctypeslib.as_array' to be a little slower than alternatives.
This eschews numpy in favor of a plain 'ctypes' solution, using 'from_address' to avoid
inducing copies.

Running the integrations/imgui.py example for 2000 iterations via timeit shows a total
runtime difference of 0.59 sec (master) vs 0.47 sec (branch) (or ~80% of the original runtime).
YMMV, but it seems generally a bit faster, and has the same phenotype as before.