Default to using imgui_bundle in the examples
Replacing pyimgui with imgui_bundle in the examples.

imgui_bundle is auto-generated, and has thus managed to follow the
original project much more closely, whereas pyimgui is manually
written and currently uses a 3.5 year old version of the Dear Imgui
C++ implementation which is lacking a lot of new features.